+++ common/newdel.cpp
@@ -57,7 +57,7 @@
  *=========================================================================*/
 void* operator new(size_t size)
 {
-    return (Alloc((unsigned long)size, MEM_NEW));
+    return (Alloc((unsigned int)size, MEM_NEW));
 }
 
 /***************************************************************************
@@ -76,7 +76,7 @@ void* operator new(size_t size)
  *=========================================================================*/
 void* operator new[](size_t size)
 {
-    return (Alloc((unsigned long)size, MEM_NEW));
+    return (Alloc((unsigned int)size, MEM_NEW));
 }
 
 /***************************************************************************
